Use superpages
The Print-a-page-table section is the most useful debugging tool.
It's worth learning conditional breakpoints and unit testing (usertests).
The most challenging issues often involve coding style (that takes me a lot of time to understand the business logic of what I have just written) and maintenance (backward compatibility in particular).